% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables_long.R
\name{convert_tabular}
\alias{convert_tabular}
\title{Convert count data to tabular form}
\usage{
convert_tabular(
  table,
  formula,
  measures = c(No. = "n"),
  ...,
  conversion = NULL,
  value_override
)
}
\arguments{
\item{table}{A data frame in semi-long format. Each dimension of the table
should have a corresponding column, but there can be multiple columns for
various measures.}

\item{formula}{A formula describing the visual structure of the table. The
LHS of the formula corresponds to rows, and the RHS corresponds to columns.
The \code{.Measure} variable can be used where there are multiple measures.
Passed to \code{data.table::melt}.}

\item{measures}{Columns that correspond to measures wanted in the output
table. The names of the vector will be used as "pretty" names in the output
table.}

\item{...}{Passed to \code{data.table::melt}.}

\item{conversion}{(Optional.) A function to convert all measures to a common
type, such as \code{as.numeric} or \code{as.character}.}

\item{value_override}{(Optional.) If a supplementary table is needed that
does not come from the measure columns (e.g. whether a cell needs to be
suppressed), this argument overrides the value used when converting to
tabular format.}
}
\value{
\code{convert_tabular} returns a \code{data.table} in visual tabular form.
Column names are determined by \code{data.table::melt}, which defaults to the
various dimensions separated by an underscore.
}
\description{
\code{convert_tabular} takes a table in semi-long format (where each row
corresponds to a cell, but each cell can have multiple measures), and returns
it in tabular format.
}
