% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables_long.R
\name{count_aggregate}
\alias{count_aggregate}
\title{Count and sum into a long table}
\usage{
count_aggregate(MT, data, ...)
}
\arguments{
\item{MT}{The \code{MappingTable} object defining the structure of the table.}

\item{data}{The raw dataset to count and aggregate.}

\item{...}{Passed to \link[dplyr:mutate-joins]{dplyr::right_join}}
}
\value{
\code{count_aggregate} returns a table of aggregated counts in long
format. The structure of the output table will be \code{MT$mtab} with an extra
column \code{n} for the counts.
}
\description{
\code{count_aggregate} counts and aggregates raw data into the structure
determined by the provided mapping table \code{MT}.
}
