% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FactorMappingTable.R
\name{FactorMappingTable}
\alias{FactorMappingTable}
\title{Basic Mapping Table for discrete categorisations}
\description{
This mapping table describes the structure of a single dimension
of a counts table, where categories are described by one or more discrete
levels.
}
\examples{
exampledata <- data.frame(groupid = rep(
  c('a', 'b', 'c', 'd', 'e', NA),
  times = c(24, 8, 12, 36, 16, 6)
))

# If we set .other = NULL, then 'e' and NA are not counted, even in Total
MT_no_other <- FactorMappingTable$new(
  'Group', 'groupid',
  'A' = 'a',
  'B' = 'b',
  'C' = 'c',
  'D' = 'd',
  'C+D' = c('c', 'd'),
  .other = NULL,
  .total = "Total"
)

MT_no_other
MT_no_other$count_aggregate(exampledata)

# Otherwise, 'e' and NA are included in both Other/Unknown and Total
MT_other <- FactorMappingTable$new(
  'Group', 'groupid',
  'A' = 'a',
  'B' = 'b',
  'C' = 'c',
  'D' = 'd',
  'C+D' = c('c', 'd'),
  .other = "Other/Unknown",
  .total = "Total"
)

MT_other
MT_other$count_aggregate(exampledata)
}
\section{Super classes}{
\code{\link[modulartabler:MappingTable]{modulartabler::MappingTable}} -> \code{\link[modulartabler:BaseMappingTable]{modulartabler::BaseMappingTable}} -> \code{FactorMappingTable}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FactorMappingTable-new}{\code{FactorMappingTable$new()}}
\item \href{#method-FactorMappingTable-preprocess}{\code{FactorMappingTable$preprocess()}}
\item \href{#method-FactorMappingTable-clone}{\code{FactorMappingTable$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="modulartabler" data-topic="MappingTable" data-id="count_aggregate"><a href='../../modulartabler/html/MappingTable.html#method-MappingTable-count_aggregate'><code>modulartabler::MappingTable$count_aggregate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="modulartabler" data-topic="MappingTable" data-id="print"><a href='../../modulartabler/html/MappingTable.html#method-MappingTable-print'><code>modulartabler::MappingTable$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FactorMappingTable-new"></a>}}
\if{latex}{\out{\hypertarget{method-FactorMappingTable-new}{}}}
\subsection{Method \code{new()}}{
Create a new mapping table object for discrete groupings.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FactorMappingTable$new(
  table_name,
  data_col,
  ...,
  .other = "Other/Unknown",
  .total = "Total"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{table_name}}{The name of the column to map to. This should generally
be human-readable.}

\item{\code{data_col}}{The name of the column in the dataset to map from.}

\item{\code{...}}{Category definitions, written as named arguments. The names
become the human-readable output categories, and the values of each
argument are the specific values in the data that all get mapped to
that category.}

\item{\code{.other}}{The name of the category to use for values not included in
\code{...}, including missing values. Set to \code{NULL} to exclude missing
values from the mapping and totals.}

\item{\code{.total}}{The name of the category to use for the total mapping,
which includes all records included in the rest of the map. Set to
\code{NULL} to omit the total category.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \link{MappingTable} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FactorMappingTable-preprocess"></a>}}
\if{latex}{\out{\hypertarget{method-FactorMappingTable-preprocess}{}}}
\subsection{Method \code{preprocess()}}{
Pre-process data for discrete mappings. For
FactorMappingTable, \code{preprocess(data)} creates a new column that maps
all variables that are not given in \code{initialize()} to a missing value.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FactorMappingTable$preprocess(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{The dataset to prepare for counting and aggregating.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FactorMappingTable-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FactorMappingTable-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FactorMappingTable$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
