% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulartabler-package.R
\docType{package}
\name{modulartabler-package}
\alias{modulartabler}
\alias{modulartabler-package}
\title{modulartabler: Building count tables from parts}
\description{
Builds tables of counts modularly. Each table is described by mapping tables for each dimension. Dimensions can be mixed and matched to build different tables with the same margins. Small cells can be suppressed, and mapping tables can be used to determine what secondary suppression is necessary to prevent back-calculation.
}
\author{
\strong{Maintainer}: Ian Powell \email{powell.ian.n@gmail.com}

Other contributors:
\itemize{
  \item Peter Moritz \email{p.a.moritz@gmail.com} [contributor]
  \item Hafiz Khusyairi [contributor]
  \item State of New South Wales NSW Ministry of Health [copyright holder]
}

}
\keyword{internal}
