% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiMappingTable.R
\name{MultiMappingTable}
\alias{MultiMappingTable}
\title{R6 class representing a multidimensional mapping table.}
\description{
R6 class representing a multidimensional mapping table.

R6 class representing a multidimensional mapping table.
}
\details{
A multidimensional mapping table is the simplest way to build a table in
multiple dimensions. It contains a cell for each level of each category in
the input tables.
}
\section{Super class}{
\code{\link[modulartabler:MappingTable]{modulartabler::MappingTable}} -> \code{MultiMappingTable}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MultiMappingTable-new}{\code{MultiMappingTable$new()}}
\item \href{#method-MultiMappingTable-preprocess}{\code{MultiMappingTable$preprocess()}}
\item \href{#method-MultiMappingTable-clone}{\code{MultiMappingTable$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="modulartabler" data-topic="MappingTable" data-id="count_aggregate"><a href='../../modulartabler/html/MappingTable.html#method-MappingTable-count_aggregate'><code>modulartabler::MappingTable$count_aggregate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="modulartabler" data-topic="MappingTable" data-id="print"><a href='../../modulartabler/html/MappingTable.html#method-MappingTable-print'><code>modulartabler::MappingTable$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultiMappingTable-new"></a>}}
\if{latex}{\out{\hypertarget{method-MultiMappingTable-new}{}}}
\subsection{Method \code{new()}}{
Create a new MultiMappingTable object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiMappingTable$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{One or more MappingTables, each representing a different
dimension of the desired table.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultiMappingTable-preprocess"></a>}}
\if{latex}{\out{\hypertarget{method-MultiMappingTable-preprocess}{}}}
\subsection{Method \code{preprocess()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiMappingTable$preprocess(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultiMappingTable-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MultiMappingTable-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiMappingTable$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
