% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_suppression.R
\name{to_dense}
\alias{to_dense}
\title{Convenience function to convert matrix to dense representation}
\usage{
to_dense(M, zapsmall = TRUE, digits = ceiling(-log10(.Machine$double.eps)/2))
}
\arguments{
\item{M}{matrix to convert to dense form}

\item{zapsmall}{Whether to zap small values to zero; see \code{?base::zapsmall}}

\item{digits}{Number of digits to zap to zero}
}
\value{
\code{to_dense} returns a matrix with three columns. The first column is
the row number, the second is the column number, and the third is the
value.
}
\description{
This function is to facilitate work with \code{lpSolve::lp}, particularly with the
\code{dense.const} argument.
}
