% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiMappingTable.R
\name{MultiMappingTable}
\alias{MultiMappingTable}
\title{R6 class representing a multidimensional mapping table.}
\description{
R6 class representing a multidimensional mapping table.

R6 class representing a multidimensional mapping table.
}
\details{
A multidimensional mapping table is the simplest way to build a table in
multiple dimensions. It contains a cell for each level of each category in
the input tables.
}
\examples{
species_table <- data.frame(
  Species = factor(
    c(1L, 2L, 3L, rep(4L, times = 3L)),
    labels = c("I. setosa", "I. versicolor", "I. virginica", "Total")
  ),
  .rawspecies = factor(rep(c("setosa", "versicolor", "virginica"), times = 2L))
)
species_table

SpeciesMap <- BaseMappingTable$new(
  species_table,
  raw_cols = ".rawspecies",
  table_cols = "Species",
  data_cols = "Species"
)
SpeciesMap

SepalMap <- RangeMappingTable$new(
  "Sepal width", 'Sepal.Width',
  "<2.5" = c(0, 2.5),
  "2.5-3.4" = c(2.5, 3.5), # sic 4.0 -- actually uses interval [2.5, 2.5)
  "3.5+" = c(3.5, Inf),
  .other = NULL, # Excludes NA and values outside range from table
  .total = "Total"
)

MultiMap <- MultiMappingTable$new(SpeciesMap, SepalMap)
MultiMap$count_aggregate(iris)
}
\section{Super class}{
\code{\link[modulartabler:MappingTable]{modulartabler::MappingTable}} -> \code{MultiMappingTable}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{map}}{The data frame used for mapping.}

\item{\code{mtab}}{The table-side data that corresponds to columns of the matrix
representation.}

\item{\code{matrix}}{A matrix representation of the mapping table that indicates
which raw values (in rows) are mapped to table cells (in columns).}

\item{\code{raw_cols}}{The names of columns in \code{map} that are joined with
preprocessed data.}

\item{\code{data_cols}}{The names of columns preprocessed data that are joined
with \code{map}.}

\item{\code{table_cols}}{The names of columns in the output dataset.}

\item{\code{join_clause}}{A \code{\link[dplyr:join_by]{dplyr::join_by()}} object that describes how to join
the data (in \code{x}) to the mapping table (in \code{y}).}

\item{\code{nullspace}}{A matrix with rowspace equal to the kernel of the matrix
representation.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MultiMappingTable-new}{\code{MultiMappingTable$new()}}
\item \href{#method-MultiMappingTable-preprocess}{\code{MultiMappingTable$preprocess()}}
\item \href{#method-MultiMappingTable-clone}{\code{MultiMappingTable$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="modulartabler" data-topic="MappingTable" data-id="count_aggregate"><a href='../../modulartabler/html/MappingTable.html#method-MappingTable-count_aggregate'><code>modulartabler::MappingTable$count_aggregate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="modulartabler" data-topic="MappingTable" data-id="print"><a href='../../modulartabler/html/MappingTable.html#method-MappingTable-print'><code>modulartabler::MappingTable$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultiMappingTable-new"></a>}}
\if{latex}{\out{\hypertarget{method-MultiMappingTable-new}{}}}
\subsection{Method \code{new()}}{
Create a new MultiMappingTable object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiMappingTable$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{One or more MappingTables, each representing a different
dimension of the desired table.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultiMappingTable-preprocess"></a>}}
\if{latex}{\out{\hypertarget{method-MultiMappingTable-preprocess}{}}}
\subsection{Method \code{preprocess()}}{
Pre-process a dataset for counting and aggregating. The
method for MultiMappingTable passes the data through the preprocessing
functions for each of its subtables.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiMappingTable$preprocess(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{The dataset to prepare for counting and aggregation.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultiMappingTable-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MultiMappingTable-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiMappingTable$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
